backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
openai
reportlab
python-dotenv
